requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
attrs==24.3.0
base58==2.1.1
bitarray==2.9.3
cached-property==2.0.1
certifi==2024.12.14
charset-normalizer==3.4.0
ckzg==2.0.1
cytoolz==0.12.3
eth-account==0.13.4
eth-bloom==3.0.1
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-tester==0.12.0b2
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
frozenlist==1.5.0
hexbytes==1.2.1
idna==3.10
ipfshttpclient==0.8.0a2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.1.0
netaddr==1.3.0
packaging==23.2
parsimonious==0.10.0
propcache==0.2.1
protobuf==3.19.5
py-ecc==7.0.1
py-evm==0.10.1b2
py-geth==5.1.0
py-solc-x==2.0.3
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pyunormalize==16.0.0
referencing==0.35.1
regex==2024.11.6
requests==2.31.0
rlp==4.0.1
rpds-py==0.22.3
safe-pysha3==1.0.4
semantic-version==2.10.0
six==1.17.0
sortedcontainers==2.4.0
toolz==1.0.0
trie==3.0.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
varint==1.0.2
web3==7.6.1
websockets==13.1
yarl==1.18.3
